this is very close to working.
The down stream messages from br.native are not getting framed
properly.

It appears that the data is transferred from br.native to slip-radio
properly --- but that either that data is incorrect (because of
something on the native br) or the netstack on slip-radio is
corrupting / mis-framing it.

The DIOs from br.native are not correct.

Correct DISs from a webserver appear to be received properly and
processeed by br.native.